Move PasswordValidator into validator package
